Incomplete Final
major apologies- there's no way I'm getting this done in time. It's partly UV mapped- just the pot and the floor- but yeah. Stuck on the UV mapping and I know the topics gonna settle in the next few weeks if I keep trying at it, but definatley not gonna happen in the next hour and a half.

on the bright side, I love modelling! That's been really fun, and I actually understand that really well!